Added 'queryGlobal' / Fixed 'reply' or 'getReply' issue with primitive values.
The 'queryGlobal' does the same as 'query', but it executes the given
javascript in the global scope, instead of the scope of a local function.
Another difference is that the excuted JS and the JS return expression are distinct.

Replies sent from JS had the problem that primitive values, like numbers or
booleans, could not be parsed by the JSON parser on server side. I fixed that
issue by adding a wrapper object for the sent data, that is unwrapped in
getReply.